Crate massa_deferred_calls

Modules

call 🔒

This module implements a new version of the Autonomous Smart Contracts. (ASC) This new version allow asynchronous calls to be registered for a specific slot and ensure his execution.

config

macros 🔒

registry_changes

slot_changes

Macros

cancelled_key

cancelled key formatting macro

coins_key

coins key formatting macro

deferred_call_field_key

key formatting macro

deferred_call_prefix_key

deferred_call_slot_base_fee_key

deferred_call_slot_total_gas_key

deferred_slot_call_prefix_key

fee_key

fee key formatting macro

max_gas_key

max gas key formatting macro

parameters_key

parameters key formatting macro

sender_address_key

sender address key formatting macro

target_address_key

target address key formatting macro

target_function_key

target function key formatting macro

target_slot_key

target slot key formatting macro

Structs

DeferredCall

Definition of a call in the future

DeferredCallRegistry

DeferredSlotCalls

Per-slot deferred-call view (calls, slot gas, base fee).

Type Aliases

DeferredRegistryBaseFeeChange

DeferredRegistryCallChange

DeferredRegistryGasChange